refactor: inline `moveRangeByOperation` into `transformRange`
`moveRangeByOperation` was a wrapper around two `transformPoint` calls plus
an identity short-circuit. Move the short-circuit into `transformRange`
itself so the engine primitive carries the "nothing changed" optimization,
then update the one call site (`range-decorations-machine`) to call
`transformRange` directly.
